addressed review comments and few amendments to the test suite
